+++ src/components/Sidebar/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./Sidebar";
+export * from "./SidebarToggleButton";
+export * from "./createLinkGroupsPlaceholder";
