--- src/pages/Dashboard/AIF/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AIF";
