--- src/pages/Dashboard/Container/SidebarToggleButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SidebarToggleButton";
