--- src/pages/Dashboard/Container/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Container";
